feat: Add comment support
Add comment parser to the grammar and simply ignore the contents during the reading phase

As is usual in Lisp, comments start with ';' and end with an end of line character (i.e. '\n', '\r')
Example: fun {fact n} {if (== n 0) {1} {* n (fact n-1) }} ; this computes the factorial

These are not of much use on the REPL, but are here now in preparation for the library system, yay!